requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
boto3==1.19.11
botocore==1.22.11
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
discord==2.0
discord-components==2.1.2
discord-py-interactions==3.0.2
discord-py-slash-command==3.0.1
discord-slash-components==0.0.2
discord.py==1.7.3
dpy-slash-button-paginator==1.0.post2
Flask==2.0.2
frozenlist==1.2.0
gifmaker==1.5
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.10.3
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.4
Pillow==8.4.0
psutil==5.8.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
requests==2.26.0
rfc3986==1.5.0
s3transfer==0.5.0
scipy==1.7.2
six==1.16.0
sniffio==1.2.0
titlecase==2.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
